fix: adjust project states to new navigation and layout
