+++ dashboard/src/features/orgs/components/common/TransferProjectDialog/index.ts
@@ -1,0 +1,1 @@
+export { default as TransferProjectDialog } from './TransferProjectDialog';
