+++ dashboard/src/features/orgs/projects/common/components/ApplicationPausedReason/ApplicationPausedReason.tsx
@@ -1,5 +1,5 @@
 import { Alert } from '@/components/ui/v2/Alert';
-import { Text } from '@/components/ui/v2/Text';
+import { useCurrentOrg } from '@/features/orgs/projects/hooks/useCurrentOrg';
 
 interface ApplicationPausedReasonProps {
   freeAndLiveProjectsNumberExceeded?: boolean;
@@ -8,23 +8,28 @@ interface ApplicationPausedReasonProps {
 export default function ApplicationPausedReason({
   freeAndLiveProjectsNumberExceeded,
 }: ApplicationPausedReasonProps) {
+  const { org } = useCurrentOrg();
+
   return (
     <Alert
       severity="warning"
-      className="mx-auto flex max-w-xs flex-col gap-4 p-6 text-left"
+      className="flex flex-col w-full max-w-xs gap-4 p-6 mx-auto text-left"
     >
-      <Text>
-        Starter projects will stop responding to API calls after 7 days of
-        inactivity, so consider
-        <span className="font-semibold"> upgrading to Pro </span>to avoid
-        auto-sleep.
-      </Text>
+      {org?.plan?.isFree ? (
+        <p>
+          Projects under your Personal Organization will stop responding to API
+          calls after 7 days of inactivity, so consider transferring the project
+          to a <b>Pro Organization</b> to avoid auto-sleep.
+        </p>
+      ) : (
+        <p className="text-center">Your project is Paused.</p>
+      )}
       {freeAndLiveProjectsNumberExceeded && (
-        <Text>
+        <p className="text-center">
           Additionally, only 1 free project can be active at any given time, so
           please pause your current active free project before unpausing
           another.
-        </Text>
+        </p>
       )}
     </Alert>
   );
